fix bug - removing spark on result
by making array type as np.float32

+++ HW2/HW2_code.py
@@ -21,7 +21,7 @@ def gauss1d(sigma):
     #generated an array of values of distance from center
     x = np.arange(len//2 * -1, len//2 + 1,1)
     #compute gaussian value
-    gaussian = np.exp(-1*x*x/(2*sigma*sigma))
+    gaussian = np.exp(-1*x*x/(2*sigma*sigma),dtype=np.float32)
     #normalize the filter so that the sum of values is 1
     return gaussian / np.sum(gaussian)
     
@@ -42,7 +42,7 @@ def convolve2d(array, filter):
     #assume that row size and column size of filter is same
 
     #make empty image
-    result = np.zeros_like(array)
+    result = np.zeros_like(array,dtype=np.float32)
     #get filter size(for calculating padding size)
     filter_row_size = filter.shape[0]
     #calculating padding size
@@ -56,7 +56,7 @@ def convolve2d(array, filter):
     #perform convolution
     for i in range(array.shape[0]):
         for j in range(array.shape[1]):
-            result[i,j] = np.sum(pad_array[i:i+filter_row_size,j:j+filter_row_size]*flip_filter)
+            result[i,j] = np.sum(pad_array[i:i+filter_row_size,j:j+filter_row_size]*flip_filter,dtype=np.float32)
 
     return result
 #b
@@ -73,12 +73,13 @@ im = Image.open('images/3b_tiger.bmp')
 #make image gray scale
 im_gray = im.convert('L')
 #convert image to array
-im_array = np.asarray(im_gray)
+im_array = np.asarray(im_gray,dtype=np.float32)
 
 #apply gaussian filter(sigma=3)
 convolved = gaussconvolve2d(im_array,3)
 # clamping the pixels on (0,255)
 convolved = np.clip(convolved, 0, 255, out=convolved)
+#convolved = convolved / convolved.max() * 255
 convolved = convolved.astype(np.uint8)
 #d
 # show original image, show filtered image, and save
@@ -100,7 +101,7 @@ def gaussconvolved2dRGB(array,sigma):
     # filter each of the three color channels (RGB) separately
     for i in range(3):
         low_frequency_result[:,:,i] = gaussconvolve2d(array[:,:,i],sigma)
-        low_frequency_result[:,:,i] = np.clip(low_frequency_result[:,:,i], 0, 255)
+        low_frequency_result[:,:,i] = np.clip(low_frequency_result[:,:,i],0,255)
     return low_frequency_result
 
 #2-1
@@ -108,12 +109,12 @@ def gaussconvolved2dRGB(array,sigma):
 im2 = Image.open('images/3b_tiger.bmp')
 
 #make array from image
-im2_array = np.asarray(im2)
+im2_array = np.asarray(im2,dtype=np.float32)
 
 low_frequency_result = gaussconvolved2dRGB(im2_array,4)
 
 # convert result array to image, show and save
-low_frequency_img = Image.fromarray(low_frequency_result)
+low_frequency_img = Image.fromarray(low_frequency_result.astype(np.uint8))
 low_frequency_img.show()
 low_frequency_img.save("part2-1_result.bmp","bmp")
 
@@ -124,7 +125,7 @@ low_frequency_img.save("part2-1_result.bmp","bmp")
 im3 = Image.open('images/3a_lion.bmp')
 
 #make array from image
-im3_array = np.asarray(im3)
+im3_array = np.asarray(im3,dtype=np.float32)
 
 #for making high frequency image, we should make low frequency image
 low_frequency = gaussconvolved2dRGB(im3_array,4)
@@ -137,7 +138,7 @@ high_frequency_result = high_frequency_result - low_frequency
 
 # convert result array to image, show and store
 #values of above high_frequency array are zero-means negative values. so we should add 128 for visualizing
-high_frequency_img = Image.fromarray(high_frequency_result + 128)
+high_frequency_img = Image.fromarray(high_frequency_result.astype(np.uint8) + 128)
 high_frequency_img.show()
 high_frequency_img.save("part2-2_result.bmp","bmp")
 
@@ -145,7 +146,7 @@ high_frequency_img.save("part2-2_result.bmp","bmp")
 #making hybrid image. hybrid image is low frequency image + high frequency image
 hybrid_array = low_frequency_result + high_frequency_result
 hybrid_array =  np.clip(hybrid_array, 0, 255)
-hybrid_image = Image.fromarray(hybrid_array)
+hybrid_image = Image.fromarray(hybrid_array.astype(np.uint8))
 hybrid_image.show()
 hybrid_image.save("part2-3_result.bmp","bmp")
